Chapter15/Exercise15.02/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1
django-configurations==2.2
djangorestframework==3.11.0
Pillow==8.0.1
plotly==4.14.1
retrying==1.3.3
six==1.15.0
pytz==2020.1
sqlparse==0.3.1
XlsxWriter==1.3.7
